Telegram_bot_Wildberries/requirements.txt
aiogram==3.0.0b7
requests==2.28.0
aiohttp==3.8.1
python-dotenv==0.21.0